+++ src/Text/Parsec/Bits.hs
@@ -1,8 +1,13 @@
 module Text.Parsec.Bits
     ( BitParser,
       BitParserT,
-      number,
-      numberFull
+      bitstream,
+      one,
+      ones,
+      ganyNumber,
+      anyNumber,
+      anyNumberFull,
+      number
     ) where
 
 import Data.Bit.Stream (BitStream, BitOrder, toBitsType)
@@ -20,6 +25,16 @@ import Data.Word (Word8)
 type BitParserT u m = ParsecT BitStream u m
 type BitParser u = BitParserT u Identity
 
+one :: BitParser a Bool
+one = tokenPrim
+        (show . fromEnum)
+        (\p _ _ -> incSourceColumn p 1)
+        (\x -> if x then Just x else Nothing)
+         <?> "a one bit"
+
+ones :: Int -> BitParser a [Bool]
+ones n = count n one <?> ("a block of " ++ (show n) ++ " one bits")
+
 bitstream :: (Monad m) => Int -> BitParserT u m BitStream
 bitstream n = do
   strm <- getInput
@@ -31,16 +46,24 @@ bitstream n = do
       setInput partB
       return partA
 
-
-gnumber :: (Monad m, Num a, Bits a) => Int -> BitOrder -> BitParserT u m a
-gnumber bitcnt order = do
+-- generalized anyNumer
+ganyNumber :: (Monad m, Num a, Bits a) => Int -> BitOrder -> BitParserT u m a
+ganyNumber bitcnt order = do
     toBitsType order <$> bitstream bitcnt
 
-number :: (Monad m) => Int -> BitOrder -> BitParserT u m Word 
-number = gnumber
+anyNumber :: (Monad m) => Int -> BitOrder -> BitParserT u m Word 
+anyNumber = ganyNumber
 
 withMonadicResultProxy :: (Proxy a -> BitParserT u m a) -> BitParserT u m a
 withMonadicResultProxy action = action undefined
 
-numberFull :: (Monad m, Num a, FiniteBits a) => BitOrder -> BitParserT u m a
-numberFull order = withMonadicResultProxy (\prx -> gnumber (finiteBitSize (zeroBits `asProxyTypeOf` prx)) order)
+anyNumberFull :: (Monad m, Num a, FiniteBits a) => BitOrder -> BitParserT u m a
+anyNumberFull order = withMonadicResultProxy (\prx -> ganyNumber (finiteBitSize (zeroBits `asProxyTypeOf` prx)) order)
+
+number :: Int -> BitOrder -> Word -> BitParser a Word
+number length order expected = flip label ("number " ++ show expected) $ try $ do
+    n <- anyNumber length order
+    if n == expected then
+        return n
+    else
+        unexpected ("number " ++ show n)