+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for bitstream
 
+Version 0.2:
+- rename number / gnumber / numberFull to anyNumber / ganyNumber / anyNumberFull
+- add one / ones
+- export bitstream
+- Add tests
 Version 0.1:
 - Initial release